Package initialization (#1)
* Package initialization

An custom implementation of equinox.nn.Linear is provided that enables custom parameter initialization. While the bias is initialized to zero by default, the weights intitialization must always be provided by the user, because not every initizialization works with every data type and every input shape. For example, He normal intialization does not accept shaped that are zero, wile uniform intialization does not work with complex numbers.

A test for the klax.nn.Linear is implementes, which is closely realted to the tests implemented in equinox.

* Updated docstrings for class Linear to account for recent changes.