Add a FullyLinear layer
The this layer is very similar to klax.nn.Linear, but receives to, instead of one input. This will be usefull for building MLPs with pass-through inputs.

Tests were added in ananoly to klax.nn.Linear with some additional test cases to account for different shapes of the two inputs.