Added parameter wrappers to `klax.nn.Linear`
To achieve a uniform interface for all parameter constraint, an abstract class `ParameterWapper` was introduced, which uses `paramax.AbstractUnwrappable`in the background.  In constrast to `paramax.AbstractUnwrappable` a class derived from `ParameterWrapper` must always receive exactly one parameter as input.

Using `ParameterWrapper` the possibility for constraints was added to the linear layer and a simple test was added, where all parameters are constrainted to zero.