Data loader for abitrary `PyTree`s
Type annotations where changed from `Sequence`to `PyTree`, because the data loader works for any `PyTree`data structure. To show this, a test with a complex nestes, data structure as added.

To make sure that the `data` and `batch_mask` arguments always have identical `PyTree`structure, an additional assert was added and tested.

Some reformulations.