Basic training loop and test
Note, that the fit function receives in_mask and out_mask as argumentes, which can be thought of as a split version of batch_mask in the dataloader. Before being passed to the dataloder, they are joined together as (in_mask, out_mask). This is more explicit than passing batch_mask directly to the fit function and involves less "hard coding".

WARNING: The test currently fails.