Fixed type annotations + typing module
All TypeAliases were moved to a new `klax.typing` module.  To create emplates for dataloaders and loss functions, protocols were created.

All type errors in the `test_training.py` are now resolved and all but one are resolved in `_training.py`. The remainig one has to to with the history.