Modified `fit` for new HistoryCallback
The funciton now receives a HistorCallback as input. If that input is
not provided, a HistoryCallback is created internally, with a input of
`log_every=100`. `log_every` was, in turn, removed from the arguments of
the function. To specify `log_every`, the user must now simply pass an
instance of `HistoryCallback` to the function, that has the desired
settings.

Logging of loss values was moved to `HistoryCallback`