Rework to partitioning and combining models
Dictinc `combined_loss` and `partitioned_loss`functions were created (basically just a renaming of the previous function). `combined_loss` is only used withing `cbargs` in order to maintain a model-based API. `partitioned_loss` is used within `make_step` where a consistent parameter-based API is used.

This ensures that the number of partitions and combinations of the model are minimized and compatibilty with all optax optimizers is maintained.