Modularized fit
The `fit` function is modualized.

The main loop is performed within the `_fit_core` function, which knows
nothing about the data, optimizer, model, and loss. It only performs the
update.

The original API of `fit` is provided as a wrapper, which exposes some
parameters of the training components to the user, e.g., batch axis.

Different `Protocols` are implemented for the loss, the updater and the
batcher, but none of them create new layers of abstraction, i.e. not new
classes, but mere serve as templates for the different functions
required for training.

Note: Callbacks have not been implemented in this approach yet.