Enable stateful batching + training state extended with treedefs
The protocol for BatchGenerators has been changed to allow for a
`TrainingState` to be sent to the generator. This enables statefull
batching, where, e.g., the step of the model can be used as a random
seed for jandom number generation.

Additionally, the task of flattening and unflattening the model was used
to the `TrainingState', making the `fit_core` function even more
minimal.