Added EvaluationContext and TrainingContext
The `EvaluationContext` takes care of the lazy loss evaluation. It
receives the training state as input and stores the loss and validation
loss in a cache. The cache is updated whenever the training step
changed.

The `TrainingContext` is a simple wrapper around a `TrainingState` and
an `EvaluatorContext`. It serves as the main input to every `Callback`.

This new implementation nicely separates the `dump` training state from
the data and the evaluation.